move PasteBuildingSettingUpdateProcessor.cs to Entity/
